wifi/mikrotik-config.tex: initial commit
I exhausted myself while writing and/or formatting this and consider using LaTeX for this a mistake, while this is probably my most fancy document yet and likely was good practice.
I learned at least that arrows require inline math and how to get table of contents that my other files don't have.